+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ Change log
 - Propagated values may now be garbage collected if their associated `Var` object goes out of scope. 
 - :func:`spox.opsets.ai.v17.loop`, :func:`spox.opsets.ai.v19.loop` and :func:`spox.opsets.ai.v21.loop` will only infer shapes for loop carried dependencies if their shapes are unchanged across iterations.
 
+
 0.13.0 (2024-12-06)
 -------------------